refactor: extract turndown updates to own file
